requirements.txt
streamlit
pandas
pydub
natsort
openpyxl
python-docx
chardet
nltk
xlsxwriter